unsetting LR since lower test acc than expected
loaded up model that should be getting 59% test acc but after just a small bit of training it's getting high 40s :-/ thought maybe manually setting the LR to a higher value (at least 1 OOM) might be causing it, but on second thought it is stable after that first bit of training. So we'll see! :)